Refactor package.json: Remove next-auth dependency
