+++ package.json
@@ -21,7 +21,6 @@
     "geist": "^1.3.0",
     "lucide-react": "^0.439.0",
     "next": "^14.2.4",
-    "next-auth": "^4.24.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "server-only": "^0.0.1",
