--- src/app/api/auth/[...nextauth]/route.ts
@@ -1,7 +1,0 @@
-import NextAuth from "next-auth";
-
-import { authOptions } from "@sdit/server/auth";
-
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };
