+++ src/app/api/trpc/[trpc]/route.ts
@@ -2,8 +2,8 @@ import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
 import { type NextRequest } from "next/server";
 
 import { env } from "@sdit/env";
-import { appRouter } from "@sdit/server/api/root";
-import { createTRPCContext } from "@sdit/server/api/trpc";
+import { appRouter } from "@sdit/server/root";
+import { createTRPCContext } from "@sdit/server/trpc";
 
 /**
  * This wraps the `createTRPCContext` helper and provides the required context for the tRPC API when
