+++ src/trpc/server.ts
@@ -4,8 +4,8 @@ import { createHydrationHelpers } from "@trpc/react-query/rsc";
 import { headers } from "next/headers";
 import { cache } from "react";
 
-import { createCaller, type AppRouter } from "@sdit/server/api/root";
-import { createTRPCContext } from "@sdit/server/api/trpc";
+import { createCaller, type AppRouter } from "@sdit/server/root";
+import { createTRPCContext } from "@sdit/server/trpc";
 import { createQueryClient } from "./query-client";
 
 /**
